Inicio das mudanças estruturais. - Tentar corrigir erro do dice_numer, dando ValueError de cara.
